+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_params.py
@@ -122,6 +122,7 @@ register_cli_argument('acs kubernetes get-credentials', 'path',
                       default=os.path.join(os.path.expanduser('~'), '.kube', 'config'),
                       type=file_type,
                       completer=FilesCompleter())
-register_cli_argument('acs scale', 'new_agent_count', type=int, help='The number of agents for the cluster')
+register_cli_argument('acs scale', 'new_agent_count', type=int, help='The number of agents for the first (default) agent pool')
+register_cli_argument('acs scale', 'new_agent_counts', type=validate_file_or_dict, help=_get_feature_in_preview_message() + 'The json map describes number of new agents in each agent pool, by pool name')
 register_cli_argument('acs create', 'service_principal', help='Service principal for making calls into Azure APIs')
 register_cli_argument('acs create', 'client_secret', help='Client secret to use with the service principal for making calls to Azure APIs')